ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.7.2
python-dotenv==1.0.1
Pillow==10.3.0
python-multipart==0.0.9
httpx==0.27.0
cachetools==5.3.3
